requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
python-decouple>=3.8
django-tailwind>=3.6.0